mod-ur: http/ws client now in rudimentary module
